Cleanup modifications to `Ppdb` and `PpdbSql` classes
This cleans up some of the modifications of the `Ppdb` and `PpdbSql`
classes so they do not include extra replica chunk information used in
the `export` module.
